requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
cvzone>=1.6.0
mysql-connector-python>=9.0.0

# Optional dependencies (requires CMake)
# face-recognition>=1.3.0
# dlib>=19.7

# Firebase (optional - for original version)
firebase-admin>=6.0.0